perf(split grid): split gird folder instead of grid files
This commit takes advantage of the newly added functions to `honeybee-radiance` to improve the performance of the recipe.

A test with a medium size model on a local computer cut the run time by 25% and decreased the number of tasks from 568 to 140 tasks! This can result in a major boost in the performance on Pollination by decreasing the need to starting up new pods.

Once/if this PR gets merged we probably want to do the same with the other recipes. This approach removes the need to input the strange `sensor_count` number and can help to reuse the recipes inside other recipes with more control on the resources.

Unfortunately, I couldn't update the annual metrics section this time but we can revise that once the new honeybee-radiance-postprocess is ready.