refactor: expose sensor count and set the default value to 1
After talking to @chriswmackey about the new strategy we decided to expose the sensor_count as a recipe input and set the default value to 1. With this design the user will have control on both parameters and by default the CPU count input will be used as the leading input unless the user changes the value for sensor count.