fix(direct): Add support for direct illuminance results next to total
We already store the direct results in the annual-irradiance recipe and I figured that we are eventually going to want the same here when we add support for ASE. Given that some advanced users were asking for these direct results and they know better that these results aren't suitable for LEED ASE, I am just adding some tasks that give them the direct result matrices. All of these tweaks are purely under the hood and I didn't change any of the outputs.